Win32 fixes and basic switchover to Qt.
 * Various fixes to newcore to make it build with Win32.
   Strange one is the max and min changes, max/min are macros in
   Visual Studio, std::min and std::max can be forced by
   specifying an explicit template parameter, otherwise wrapping
   it in '(' and ')' fixes it.
   Also, had to implement operator< for Vec2, for the 'min'
   use, not sure why it works on other systems, OpenEXR is the
   same.
 * All mouse interaction is now working as with FLTK version.
 * Keyboard captured, ESC exits.

+++ prototypes/newcore/api.cpp
@@ -941,7 +941,7 @@ RtVoid RenderApi::Option(RtConstToken name, const ParamList& pList)
                         Ri::TypeSpec(Ri::TypeSpec::Int, 2), "bucketsize"))
             m_opts->opts->bucketSize = Vec2(bs[0], bs[1]);
         if(IntArray gs = params.findIntData(Ri::TypeSpec::Int, "gridsize"))
-            m_opts->opts->gridSize = ifloor(sqrt(gs[0]));
+            m_opts->opts->gridSize = ifloor(sqrt((float)gs[0]));
         if(IntArray es = params.findIntData(Ri::TypeSpec::Int, "eyesplits"))
             m_opts->opts->eyeSplits = es[0];
         if(IntArray t = params.findIntData(Ri::TypeSpec::Int, "threads"))
